Create ChoosePlayerList.jsx
Created a component for the list of players to choose from to assign teams for easier code maintainability.
